// src/common/constants/user-roles.enum.ts
export enum EUserRoles {
  COUNTRY_CREATE = 'COUNTRY_CREATE',
  COUNTRY_DELETE = 'COUNTRY_DELETE',
  COUNTRY_UPDATE = 'COUNTRY_UPDATE',

  LEAGUE_CREATE = 'LEAGUE_CREATE',
  LEAGUE_DELETE = 'LEAGUE_DELETE',
  LEAGUE_UPDATE = 'LEAGUE_UPDATE',

  // using in classic match, data from real world
  REAL_TEAM_CREATE = 'REAL_TEAM_CREATE',
  REAL_TEAM_UPDATE = 'REAL_TEAM_UPDATE',
  REAL_TEAM_DELETE = 'REAL_TEAM_DELETE',

  REAL_PLAYER_CREATE = 'REAL_PLAYER_CREATE',
  REAL_PLAYER_UPDATE = 'REAL_PLAYER_UPDATE',
  REAL_PLAYER_DELETE = 'REAL_PLAYER_DELETE',
}